refactor(compiler-wasm): split wasm project.rs into domain modules
Pure code move mirroring the native crate split. The 2290-line project.rs
becomes a 155-line parent plus project/{config,build_info,files,css,recipes,
codegen,introspect,transforms,serde_types,support}.rs.
